backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-CORS
Flask-Limiter
email-validator
python-dotenv
requests
gunicorn
psycopg2-binary